RCJ-46 Move UserProfileService to core module
